Introduce graal-tools for Graal-specific code
Also makes it compile-time only and moves the code into its own package.
