+++ build.sbt
@@ -112,6 +112,7 @@ lazy val root = (project in file("."))
     `akka-client`,
     operator,
     `tck`,
+    `graal-tools`,
     docs
   )
   .settings(common)
@@ -404,6 +405,7 @@ def sharedNativeImageSettings(targetDir: File) = Seq(
 
 lazy val `proxy-core` = (project in file("proxy/core"))
   .enablePlugins(DockerPlugin, AkkaGrpcPlugin, JavaAgent, AssemblyPlugin, GraalVMPlugin, BuildInfoPlugin)
+  .dependsOn(`graal-tools` % Provided) // Only needed for compilation
   .settings(
     common,
     name := "cloudstate-proxy-core",
@@ -415,12 +417,6 @@ lazy val `proxy-core` = (project in file("proxy/core"))
         "io.grpc" % "grpc-netty-shaded" % GrpcJavaVersion,
         // Since we exclude Aeron, we also exclude its transitive Agrona dependency, so we need to manually add it HERE
         "org.agrona" % "agrona" % "0.9.29",
-        "org.graalvm.nativeimage" % "svm" % GraalVersion % "provided", // Only needed for compilation
-
-        // Adds configuration to let Graal Native Image (SubstrateVM) work
-        "com.github.vmencik" %% "graal-akka-actor" % GraalAkkaVersion % "provided", // Only needed for compilation
-        "com.github.vmencik" %% "graal-akka-stream" % GraalAkkaVersion % "provided", // Only needed for compilation
-        "com.github.vmencik" %% "graal-akka-http" % GraalAkkaVersion % "provided", // Only needed for compilation
         "com.typesafe.akka" %% "akka-remote" % AkkaVersion excludeAll (excludeTheseDependencies: _*),
         // For Eventing support of Google Pubsub
         "com.google.api.grpc" % "grpc-google-cloud-pubsub-v1" % "0.12.0" % "protobuf", // ApacheV2
@@ -503,13 +499,7 @@ lazy val `proxy-cassandra` = (project in file("proxy/cassandra"))
         "com.typesafe.akka" %% "akka-persistence-cassandra" % AkkaPersistenceCassandraVersion excludeAll (
           (excludeTheseDependencies :+ ExclusionRule("com.github.jnr")): _* // Can't native-image this, so we don't need this either
         ),
-        "com.typesafe.akka" %% "akka-persistence-cassandra-launcher" % AkkaPersistenceCassandraVersion % Test,
-        "org.graalvm.nativeimage" % "svm" % GraalVersion % "provided", // Only needed for compilation
-
-        // Adds configuration to let Graal Native Image (SubstrateVM) work
-        "com.github.vmencik" %% "graal-akka-actor" % GraalAkkaVersion % "provided", // Only needed for compilation
-        "com.github.vmencik" %% "graal-akka-stream" % GraalAkkaVersion % "provided", // Only needed for compilation
-        "com.github.vmencik" %% "graal-akka-http" % GraalAkkaVersion % "provided" // Only needed for compilation
+        "com.typesafe.akka" %% "akka-persistence-cassandra-launcher" % AkkaPersistenceCassandraVersion % Test
       ),
     fork in run := true,
     mainClass in Compile := Some("io.cloudstate.proxy.CloudStateProxyMain"),
@@ -538,13 +528,7 @@ lazy val `proxy-postgres` = (project in file("proxy/postgres"))
     common,
     name := "cloudstate-proxy-postgres",
     libraryDependencies ++= Seq(
-        "org.postgresql" % "postgresql" % "42.2.6",
-        "org.graalvm.nativeimage" % "svm" % GraalVersion % "provided", // Only needed for compilation
-
-        // Adds configuration to let Graal Native Image (SubstrateVM) work
-        "com.github.vmencik" %% "graal-akka-actor" % GraalAkkaVersion % "provided", // Only needed for compilation
-        "com.github.vmencik" %% "graal-akka-stream" % GraalAkkaVersion % "provided", // Only needed for compilation
-        "com.github.vmencik" %% "graal-akka-http" % GraalAkkaVersion % "provided" // Only needed for compilation
+        "org.postgresql" % "postgresql" % "42.2.6"
       ),
     fork in run := true,
     mainClass in Compile := Some("io.cloudstate.proxy.jdbc.CloudStateJdbcProxyMain"),
@@ -906,6 +890,22 @@ lazy val `tck` = (project in file("tck"))
     executeTests in IntegrationTest := (executeTests in IntegrationTest)
         .dependsOn(`proxy-core` / assembly, `java-shopping-cart` / assembly, `scala-shopping-cart` / assembly)
         .value
+  )
+
+lazy val `graal-tools` = (project in file("graal-tools"))
+  .enablePlugins(GraalVMPlugin)
+  .settings(
+    libraryDependencies ++= List(
+        "org.graalvm.nativeimage" % "svm" % GraalVersion,
+        // Adds configuration to let Graal Native Image (SubstrateVM) work
+        "com.github.vmencik" %% "graal-akka-actor" % GraalAkkaVersion,
+        "com.github.vmencik" %% "graal-akka-stream" % GraalAkkaVersion,
+        "com.github.vmencik" %% "graal-akka-http" % GraalAkkaVersion,
+        "com.typesafe.akka" %% "akka-actor" % AkkaVersion,
+        "com.typesafe.akka" %% "akka-protobuf" % AkkaVersion,
+        "com.google.protobuf" % "protobuf-java" % ProtobufVersion,
+        "com.thesamet.scalapb" %% "scalapb-runtime" % scalapb.compiler.Version.scalapbVersion
+      )
   )
 
 def doCompileK8sDescriptors(dir: File,
