+++ graal-tools/src/main/java/io/cloudstate/graaltools/Substitutions.java
@@ -1,10 +1,46 @@
 // FIXME THIS FILE IS FOR PROOF-OF-CONCEPT ONLY
-// THESE SUBSTITUTIONS SHOULD BE PRESENT IN THE SHADED VERSION OF NETTY IN GRPC-NETTY-SHADED,
-// AGRONA, AND IN PROTOBUF_JAVA. NOT HERE!
+// THESE SUBSTITUTIONS SHOULD BE PRESENT IN NETTY, IN GRPC-NETTY-SHADED, AGRONA, AND IN
+// PROTOBUF_JAVA. NOT HERE!
 
-package io.cloudstate.internal.svm;
+package io.cloudstate.graaltools;
 
 import com.oracle.svm.core.annotate.*;
+
+@TargetClass(
+    className = "io.netty.util.internal.shaded.org.jctools.util.UnsafeRefArrayAccess",
+    onlyWith = Existence.class)
+final class Target_io_netty_util_internal_shaded_org_jctools_util_UnsafeRefArrayAccess {
+  @Alias
+  @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.ArrayIndexShift, declClass = Object[].class)
+  public static int REF_ELEMENT_SHIFT;
+}
+
+@TargetClass(className = "io.netty.util.internal.CleanerJava6", onlyWith = Existence.class)
+final class Target_io_netty_util_internal_CleanerJava6 {
+  @Alias
+  @RecomputeFieldValue(
+      kind = RecomputeFieldValue.Kind.FieldOffset,
+      declClassName = "java.nio.DirectByteBuffer",
+      name = "cleaner")
+  private static long CLEANER_FIELD_OFFSET;
+}
+
+@TargetClass(className = "io.netty.util.internal.PlatformDependent", onlyWith = Existence.class)
+final class Target_io_netty_util_internal_PlatformDependent {
+  @Alias
+  @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.ArrayBaseOffset, declClass = byte[].class)
+  private static long ARRAY_BASE_OFFSET;
+}
+
+@TargetClass(className = "io.netty.util.internal.PlatformDependent0", onlyWith = Existence.class)
+final class Target_io_netty_util_internal_PlatformDependent0 {
+  @Alias
+  @RecomputeFieldValue(
+      kind = RecomputeFieldValue.Kind.FieldOffset,
+      declClassName = "java.nio.Buffer",
+      name = "address")
+  private static long ADDRESS_FIELD_OFFSET;
+}
 
 @TargetClass(
     className =
