+++ proxy/core/src/graal/META-INF/native-image/com.typesafe.akka/akka-cluster-tools/reflect-config.json.conf
@@ -17,8 +17,4 @@
 {
   name: "akka.cluster.singleton.ClusterSingletonMessage"
 }
-{
-  name: "akka.cluster.singleton.ClusterSingletonProxy"
-  allDeclaredFields: true
-}
 ]
