+++ proxy/core/src/graal/META-INF/native-image/com.typesafe.akka/akka-remote/reflect-config.json.conf
@@ -12,11 +12,6 @@
   allDeclaredConstructors: true
 }
 {
-  name: "akka.remote.RemoteDeploymentWatcher"
-  allDeclaredFields: true
-  methods: [{name: "<init>", parameterTypes: []}]
-}
-{
   name: "akka.remote.RemoteScope"
 }
 {
